Implement multithreaded back end compression by splitting up the compression stream into multiple threads, dependant on the number of CPUs detected. This facilitates a massive speed up on SMP machines proportional to the number of CPUs during the back end compression phase, but does so at some cost to the final size. Limit the number of threads to ensure that each thread at least works on a window of STREAM_BUFSIZE. Disable the lzma threading library as it does not contribute any more to the scalability of this new approach, yet compromises compression. Increase the size of the windows passed to all the compression back end types now as they need more to split them up into multiple threads, and the number of blocks increases the compressed size slightly.

+++ rzip.h
@@ -35,6 +35,8 @@
 #include <signal.h>
 #include <bzlib.h>
 #include <zlib.h>
+#include <pthread.h>
+#include <semaphore.h>
 #include <sys/resource.h>
 #include <netinet/in.h>
 
@@ -233,7 +235,7 @@ struct rzip_control {
 	unsigned long long window;
 	unsigned long flags;
 	unsigned long long ramsize;
-	unsigned long threads;
+	int threads;
 	int nice_val;		// added for consistency
 	int major_version;
 	int minor_version;
